requirements.txt
locustio==0.8.1
click==6.7
